Add CRUD comment functionality in detail page

+++ movie-app/src/app/movies/details/details.component.ts
@@ -6,23 +6,28 @@ import { ApiService } from '../../services/api.service';
 import { SplitStringPipe } from '../../shared/pipes/split-string.pipe';
 import { UserService } from '../../services/user.service';
 import { User } from '../../types/user';
+import { FormsModule, NgForm } from '@angular/forms';
 
 @Component({
   selector: 'app-details',
   standalone: true,
-  imports: [YouTubePlayerModule, SplitStringPipe, RouterLink],
+  imports: [YouTubePlayerModule, SplitStringPipe, RouterLink, FormsModule],
   templateUrl: './details.component.html',
   styleUrl: './details.component.css'
 })
 export class DetailsComponent implements OnInit {
-  movie = {} as Movie;
+  private apiLoaded = false;
   user = {} as User | null;
+
+  movie = {} as Movie;
   movieId: string = '';
 
   //TODO remove Any type
   comments: any = [];
-  
-  private apiLoaded = false;
+  isWritingComment = false;
+  isEditingComment = false;
+  editingCommentId: string | null = null;
+  editingCommentContent: string | null = null;
 
   constructor(private activatedRoute: ActivatedRoute, private apiService: ApiService, private userService: UserService, private router: Router) { }
 
@@ -30,7 +35,7 @@ export class DetailsComponent implements OnInit {
     const user = this.userService.user$.subscribe((userData) => {
       this.user = userData;
     })
-    
+
     const id = this.activatedRoute.snapshot.params['movieId'];
     this.movieId = id;
 
@@ -48,7 +53,6 @@ export class DetailsComponent implements OnInit {
     this.apiService.getAllComments(this.movieId).subscribe({
       next: (commentsData) => {
         this.comments = commentsData;
-        console.log(this.comments);
       },
       error: (err) => {
         console.error('Error fetching comments:', err);
@@ -57,8 +61,73 @@ export class DetailsComponent implements OnInit {
   };
 
   deleteMovie() {
+    const confirmation = confirm(`Do you want to delete this movie?`)
+    if (!confirmation) {
+        return;
+    }
+
     this.apiService.deleteMovie(this.movieId).subscribe(() => {
       this.router.navigate(['/catalog'])
     })
   };
+
+  addCommentToggle() {
+    this.isWritingComment = !this.isWritingComment
+  }
+
+  editCommentToggle(commentId: string, commentContent: string) {
+    this.isEditingComment = !this.isEditingComment;
+
+    if (this.isEditingComment) {
+      this.editingCommentId = commentId;
+      this.editingCommentContent = commentContent;
+    } else {
+      this.editingCommentId = null;
+      this.editingCommentContent = null;
+    }
+  }
+
+  addComment(form: NgForm) {
+    if (form.invalid) {
+      return;
+    }
+
+    this.apiService.createComment(this.movieId, form.value.comment).subscribe(() => {
+      this.apiService.getAllComments(this.movieId).subscribe((commentsData) => {
+        this.comments = commentsData;
+      });
+
+      this.addCommentToggle();
+    });
+  }
+
+  updateComment(editForm: NgForm) {
+    if (editForm.invalid || !this.editingCommentId) {
+      return;
+    }
+
+    const updatedContent = editForm.value.editComment;
+
+    this.apiService.updateComment(this.editingCommentId, updatedContent).subscribe(() => {
+      this.apiService.getAllComments(this.movieId).subscribe((commentsData) => {
+        this.comments = commentsData;
+      });
+
+      this.editCommentToggle('', '');
+    });
+  }
+
+  deleteComment(commentId: string) {
+    const confirmation = confirm(`Do you want to delete this comment?`)
+    if (!confirmation) {
+        return;
+    }
+
+    this.apiService.deleteComment(commentId).subscribe(() => {
+      this.apiService.getAllComments(this.movieId).subscribe((commentsData) => {
+        this.comments = commentsData;
+      });
+
+    })
+  }
 }
